EnergyCampus: refactor
- use new ISystem interface to move whole CityGML implementation into CityGMLSystem
- move getCSVStreams to coReadCSVUtil
- moved struct SolarPanelConfig from EnergyPlugin to SolarPanel.h
- moved struct StaticPowerData and StaticPowerCampusData to CityGMLSystem
