+++ src/OpenCOVER/plugins/hlrs/Energy/app/presentation/SolarPanel.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <lib/core/interfaces/ISolarPanel.h>
+#include <lib/core/utils/osgUtils.h>
 
 #include <osg/CopyOp>
 #include <osg/Geode>
@@ -7,19 +8,21 @@
 #include <osgDB/Options>
 
 struct SolarPanelConfig {
-  SolarPanelConfig(const std::string &name, const osg::Vec3 &position,
-                   osg::ref_ptr<osg::Group> parent, osg::ref_ptr<osg::Node> geo)
-      : name(name), position(position), parent(parent), solarPanelNode(geo) {}
-
   std::string name;
-  osg::Vec3 position;
+  float zOffset;
+  float numMaxPanels;
+  float panelWidth;
+  float panelHeight;
+  osg::Vec4 colorIntensity;
+  osg::Matrixd rotation;
   osg::ref_ptr<osg::Group> parent;
-  osg::ref_ptr<osg::Node> solarPanelNode;
+  osg::ref_ptr<osg::Geode> geode;
+  std::vector<core::utils::osgUtils::instancing::GeometryData> masterGeometryData;
+  bool valid() const { return parent && geode && !masterGeometryData.empty(); }
 };
 
 class SolarPanel : public core::interface::ISolarPanel {
  public:
-  //   SolarPanel(const SolarPanelConfig &config);
   SolarPanel(osg::ref_ptr<osg::Node> node) : m_node(node) { init(); }
 
   ~SolarPanel() {};
@@ -29,6 +32,5 @@ class SolarPanel : public core::interface::ISolarPanel {
 
  private:
   void init();
-  //   SolarPanelConfig m_config;
   osg::ref_ptr<osg::Node> m_node;
 };
